+++ resources/config_session.rb
@@ -1,8 +1,8 @@
 #
-# Cookbook Name:: grafana
+# Cookbook:: grafana
 # Resource:: config_session
 #
-# Copyright 2018, Sous Chefs
+# Copyright:: 2018, Sous Chefs
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -41,7 +41,7 @@ action :install do
 
       variables['grafana']['session'] ||= {}
       variables['grafana']['session']['provider'] ||= '' unless new_resource.session_provider.nil?
-      variables['grafana']['session']['provider'] << new_resource.provider.to_s unless new_resource.session_provider.nil?
+      variables['grafana']['session']['provider'] << new_resource.session_provider.to_s unless new_resource.session_provider.nil?
       variables['grafana']['session']['provider_config'] ||= '' unless new_resource.provider_config.nil?
       variables['grafana']['session']['provider_config'] << new_resource.provider_config.to_s unless new_resource.provider_config.nil?
       variables['grafana']['session']['cookie_name'] ||= '' unless new_resource.cookie_name.nil?